ai_agent/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
lightgbm
scipy
python-dotenv
openai
python-multipart